data-plane/shared/routing-proxy/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
psycopg2-binary>=2.9.9
